Only reply if message begins with bot name
This avoids when the bot is mentioned during a conversation, but the
user isn't trying to issue a command. For example, if the bot gets a
karma boost (bot_name++), the bot thinks there's a command being issued
to it, and the HAL message shows up. We don't want this behaviour.